tests/requirements.txt
attrs==19.3.0
importlib-metadata==1.6.0
intervaltree==3.0.2
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
PyMuPDF==1.16.18
pyparsing==2.4.7
pytest==5.4.2
PyYAML==5.3.1
six==1.14.0
sortedcontainers==2.1.0
titlecase==0.12.0
wcwidth==0.1.9
zipp==3.1.0
